chore: cleanup dead code + sizes
